Wired new counter, not running because of doubles
The two function blocks are wired up appropriately, 
but the model doesn't run because they input or 
output doubles, which can't be synthesized.

The solution is to use the fixed point design object 
constructors fi(), fimath(), or numerictype().
Examples are available by running this command 
in matlab:
open_system('eml_hdl_design_patterns')